+++ buildpacks/php/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Drop use of Classic buildpack repository (for Composer installer plugin and web server boot scripts and configs) during platform packages installation
+- Output filtered, "human-readable" progress info during platform packages installation
 
 ## [1.0.10] - 2025-10-24